+++ src/test/java/kmlib/starsector/settings/modmanager/ModPresenceTest.java
@@ -1,4 +1,4 @@
-package kmlib.starsector.settings;
+package kmlib.starsector.settings.modmanager;
 
 import kmlib.testfixtures.starsector.settings.ModStateScopes;